+++ lib/prometheus/Prometheus.h
@@ -19,7 +19,7 @@ class Prometheus
 {
 public:
 	Prometheus(String name, float value, String type, String description);
-	Prometheus(String name, float value, String type, String description, uint16_t timestamp);
+	Prometheus(String name, float value, String type, String description, uint32_t timestamp);
 	void attribute(String name, String value);
 	String value_type();
 	String help();
@@ -32,7 +32,7 @@ private:
 	String type;
 	String description;
 	float value;
-	uint16_t timestamp = 0;
+	uint32_t timestamp = 0;
 };
 
 #endif /* SRC_UI_PROMETHEUS_H_ */
